+++ components/Feature/ReferralForm/index.js
@@ -1,0 +1,643 @@
+import React from 'react';
+import { useState } from 'react';
+import useReferral from 'lib/api/utils/useReferral';
+import { sendDataToAnalytics, getUserGroup } from 'lib/utils/analytics';
+import { REFERRAL_SUBMIT_SUCCESS, REFERRAL_SUBMIT_INVALID } from 'lib/utils/analyticsConstants';
+
+const ReferralForm = ({
+  residentInfoCallback,
+  token,
+  setReferralCompletion,
+  referralCompletion,
+  referralSummary,
+  setReferralSummary,
+  updateEmailBody,
+  setEmailBody,
+  referrerData,
+  id,
+  email,
+  name,
+  description,
+  websites,
+  address,
+  telephone,
+  referralContact,
+  referralData
+}) => {
+  const { createReferral } = useReferral({ token });
+  const [residentInfo, setResidentInfo] = useState({
+    name: null,
+    email: null,
+    phone: null,
+    address: null,
+    postcode: null
+  });
+
+  const [validationError, setValidationError] = useState({});
+
+  const handleOnChange = (id, value) => {
+    delete validationError[id];
+    let newResidentInfo = { ...residentInfo, [id]: value };
+    setResidentInfo(newResidentInfo);
+    residentInfoCallback(newResidentInfo);
+  };
+
+  const onInvalidField = value => {
+    setValidationError(x => {
+      return { [value]: true, ...x };
+    });
+  };
+
+  const onSubmitForm = async e => {
+    e.preventDefault();
+    e.target.Submit.setAttribute('disabled', true);
+    const serviceId = e.target['service-id'].value;
+    const serviceName = e.target['service-name'].value;
+    const serviceTelephone = e.target['service-contact-phone'].value;
+    const serviceEmail = e.target['service-contact-email'].value;
+    const serviceAddress = e.target['service-address'].value;
+    const serviceWebsites = e.target['service-websites'].value;
+    const referralEmail = e.target['service-referral-email'].value;
+
+    const referral = {
+      firstName: e.target.firstName.value,
+      lastName: e.target.lastName.value,
+      phone: e.target.phone.value,
+      email: e.target.email.value,
+      address: e.target.address.value,
+      postcode: e.target.postcode.value,
+      referralReason: e.target['referral-reason'].value,
+      conversationNotes: e.target['conversation-notes'].value,
+      referrerName: e.target['referer-name'].value,
+      referrerEmail: e.target['referer-email'].value,
+      referrerOrganisation: e.target['referer-organisation'].value,
+      dateOfBirth: {
+        year: e.target['date-of-birth-year'].value,
+        month: e.target['date-of-birth-month'].value,
+        day: e.target['date-of-birth-day'].value
+      },
+      serviceId,
+      serviceName,
+      serviceContactEmail: e.target['service-contact-email'].value,
+      serviceReferralEmail: e.target['service-referral-email'].value,
+      serviceContactPhone: e.target['service-contact-phone'].value,
+      serviceAddress: e.target['service-address'].value,
+      serviceWebsite: e.target['service-websites'].value,
+      serviceDescription: e.target['service-description'].value,
+      sendResidentSms: e.target['resident-referral-sms'].checked,
+      sendResidentEmail: e.target['resident-referral-email'].checked
+    };
+    const result = await createReferral(referral);
+
+    if (result.id) {
+      setReferralCompletion({ ...referralCompletion, [serviceId]: result });
+      const newReferralSummary = referralSummary.concat([
+        {
+          name: serviceName,
+          telephone: serviceTelephone,
+          contactEmail: serviceEmail,
+          referralEmail,
+          address: serviceAddress,
+          websites: serviceWebsites,
+          id: serviceId,
+          referralId: result.id
+        }
+      ]);
+      setReferralSummary(newReferralSummary);
+      setEmailBody(updateEmailBody(undefined, newReferralSummary));
+      sendDataToAnalytics({
+        action: getUserGroup(referrerData['user-groups']),
+        category: REFERRAL_SUBMIT_SUCCESS,
+        label: serviceName
+      });
+    }
+  };
+  return (
+    <div>
+      <h3 className="govuk-heading-m" id="resident-details-header">
+        Who are you helping?
+      </h3>
+
+      {
+        <form
+          id={`referral-form-${id}`}
+          onInvalid={() => {
+            sendDataToAnalytics({
+              action: getUserGroup(referrerData['user-groups']),
+              category: REFERRAL_SUBMIT_INVALID
+            });
+          }}
+          onSubmit={onSubmitForm}>
+          <div
+            className={`govuk-form-group ${
+              validationError.firstName ? 'govuk-form-group--error' : ''
+            }`}>
+            <div className="govuk-!-padding-bottom-2">
+              <label className="govuk-label inline" htmlFor="firstName">
+                First name
+              </label>
+              <span id="name-error" className="govuk-error-message">
+                <span hidden={!validationError.firstName} data-testid="name-error">
+                  Enter the first name
+                </span>
+              </span>
+              <input
+                className={`govuk-input  ${validationError.firstName ? 'govuk-input--error' : ''}`}
+                id="firstName"
+                name="firstName"
+                type="text"
+                onChange={e => handleOnChange(e.target.id, e.target.value)}
+                aria-describedby="firstName"
+                onInvalid={e => onInvalidField(e.target.id)}
+                required
+              />
+            </div>
+          </div>
+          <div
+            className={`govuk-form-group ${
+              validationError.lastName ? 'govuk-form-group--error' : ''
+            }`}>
+            <div className="govuk-!-padding-bottom-2">
+              <label className="govuk-label inline" htmlFor="lastName">
+                Last name
+              </label>
+              <span id="name-error" className="govuk-error-message">
+                <span hidden={!validationError.lastName} data-testid="name-error">
+                  Enter the last name
+                </span>
+              </span>
+              <input
+                className={`govuk-input  ${validationError.lastName ? 'govuk-input--error' : ''}`}
+                id="lastName"
+                name="lastName"
+                type="text"
+                aria-describedby="lastName"
+                onChange={e => handleOnChange(e.target.id, e.target.value)}
+                onInvalid={e => onInvalidField(e.target.id)}
+                required
+              />
+            </div>
+          </div>
+          <div
+            className={`govuk-form-group ${
+              validationError.phone ? 'govuk-form-group--error' : ''
+            }`}>
+            <div className="govuk-!-padding-bottom-2">
+              <label className="govuk-label inline" htmlFor="phone">
+                Phone
+              </label>
+              <span id="phone-error" className="govuk-error-message">
+                <span hidden={!validationError.phone} data-testid="phone-error">
+                  Enter the phone number
+                </span>
+              </span>
+              <input
+                className={`govuk-input  ${validationError.phone ? 'govuk-input--error' : ''}`}
+                id="phone"
+                name="phone"
+                type="number"
+                onChange={e => handleOnChange(e.target.id, e.target.value)}
+                required
+                onInvalid={e => onInvalidField(e.target.id)}
+              />
+            </div>
+          </div>
+          <div
+            className={`govuk-form-group ${
+              validationError.email ? 'govuk-form-group--error' : ''
+            }`}>
+            <div className="govuk-!-padding-bottom-2">
+              <label className="govuk-label inline" htmlFor="email">
+                Email
+              </label>
+              <span id="email-error" className="govuk-error-message">
+                <span hidden={!validationError.email} data-testid="email-error">
+                  Enter the email address
+                </span>
+              </span>
+              <input
+                className={`govuk-input  ${validationError.email ? 'govuk-input--error' : ''}`}
+                id="email"
+                name="email"
+                type="email"
+                spellCheck="false"
+                onChange={e => handleOnChange(e.target.id, e.target.value)}
+                required
+                onInvalid={e => onInvalidField(e.target.id)}
+              />
+            </div>
+          </div>
+          <div
+            className={`govuk-form-group ${
+              validationError.address ? 'govuk-form-group--error' : ''
+            }`}>
+            <div className="govuk-!-padding-bottom-2">
+              <label className="govuk-label inline" htmlFor="address">
+                Address
+              </label>
+              <span id="address-error" className="govuk-error-message">
+                <span hidden={!validationError.address} data-testid="address-error">
+                  Enter the address
+                </span>
+              </span>
+              <input
+                className={`govuk-input  ${validationError.address ? 'govuk-input--error' : ''}`}
+                id="address"
+                name="event-name"
+                type="text"
+                onChange={e => handleOnChange(e.target.id, e.target.value)}
+                required
+                onInvalid={e => onInvalidField(e.target.id)}
+              />
+            </div>
+          </div>
+          <div
+            className={`govuk-form-group ${
+              validationError.postcode ? 'govuk-form-group--error' : ''
+            }`}>
+            <div className="govuk-!-padding-bottom-2">
+              <label className="govuk-label inline" htmlFor="postcode">
+                Postcode
+              </label>
+              <span id="postcode-error" className="govuk-error-message">
+                <span hidden={!validationError.postcode} data-testid="postcode-error">
+                  Enter the postcode
+                </span>
+              </span>
+              <input
+                className={`govuk-input  ${validationError.postcode ? 'govuk-input--error' : ''}`}
+                id="postcode"
+                name="event-name"
+                type="text"
+                onChange={e => handleOnChange(e.target.id, e.target.value)}
+                required
+                onInvalid={e => onInvalidField(e.target.id)}
+              />
+            </div>
+          </div>
+          <div className="govuk-form-group">
+            <fieldset className="govuk-fieldset" role="group">
+              <legend className="govuk-fieldset__legend">Date of birth</legend>
+              <div className="govuk-date-input" id="date-of-birth">
+                <div className="govuk-date-input__item">
+                  <div className="govuk-form-group">
+                    <label
+                      className="govuk-label govuk-date-input__label"
+                      htmlFor="date-of-birth-day">
+                      Day
+                    </label>
+                    <input
+                      className="govuk-input govuk-date-input__input govuk-input--width-2"
+                      id="date-of-birth-day"
+                      name="date-of-birth-day"
+                      type="text"
+                      pattern="[0-9]*"
+                      inputMode="numeric"
+                      onChange={e => handleOnChange(e.target.id, e.target.value)}
+                    />
+                  </div>
+                </div>
+                <div className="govuk-date-input__item">
+                  <div className="govuk-form-group">
+                    <label
+                      className="govuk-label govuk-date-input__label"
+                      htmlFor="date-of-birth-month">
+                      Month
+                    </label>
+                    <input
+                      className="govuk-input govuk-date-input__input govuk-input--width-2"
+                      id="date-of-birth-month"
+                      name="date-of-birth-month"
+                      type="text"
+                      pattern="[0-9]*"
+                      inputMode="numeric"
+                      onChange={e => handleOnChange(e.target.id, e.target.value)}
+                    />
+                  </div>
+                </div>
+                <div className="govuk-date-input__item">
+                  <div className="govuk-form-group">
+                    <label
+                      className="govuk-label govuk-date-input__label"
+                      htmlFor="date-of-birth-year">
+                      Year
+                    </label>
+                    <input
+                      className="govuk-input govuk-date-input__input govuk-input--width-4"
+                      id="date-of-birth-year"
+                      name="date-of-birth-year"
+                      type="text"
+                      pattern="[0-9]*"
+                      inputMode="numeric"
+                      onChange={e => handleOnChange(e.target.id, e.target.value)}
+                    />
+                  </div>
+                </div>
+              </div>
+            </fieldset>
+            <input id={`service-name-${id}`} name="service-name" value={name} type="text" hidden />
+            <input
+              id={`service-contact-email-${id}`}
+              name="service-contact-email"
+              value={email}
+              type="text"
+              hidden
+            />
+            <input
+              id={`service-contact-phone-${id}`}
+              name="service-contact-phone"
+              value={telephone}
+              type="text"
+              hidden
+            />
+            <input
+              id={`service-referral-email-${id}`}
+              name="service-referral-email"
+              value={referralContact}
+              type="text"
+              hidden
+            />
+            <input
+              id={`service-address-${id}`}
+              name="service-address"
+              value={address}
+              type="text"
+              hidden
+            />
+            <input
+              id={`service-websites-${id}`}
+              name="service-websites"
+              value={websites.join(' ')}
+              type="text"
+              hidden
+            />
+            <input
+              id={`service-description-${id}`}
+              name="service-description"
+              value={description}
+              type="text"
+              hidden
+            />
+            <div
+              className={`govuk-form-group govuk-!-padding-bottom-2 govuk-!-padding-top-4 ${
+                validationError[`referral-reason-${id}`] ? 'govuk-form-group--error' : ''
+              }`}>
+              <legend className="govuk-fieldset__legend govuk-fieldset__legend--s">
+                Reason for referral, please give as much detail as possible
+              </legend>
+              {validationError[`referral-reason-${id}`] && (
+                <span
+                  id="more-detail-error"
+                  className={` ${
+                    validationError[`referral-reason-${id}`] ? 'govuk-error-message' : ''
+                  }`}>
+                  <span className="govuk-visually-hidden">Error:</span> Enter more detail
+                </span>
+              )}
+              <textarea
+                className={`govuk-textarea ${
+                  validationError[`referral-reason-${id}`] ? 'govuk-input--error' : ''
+                }`}
+                id={`referral-reason-${id}`}
+                name="referral-reason"
+                rows="5"
+                aria-describedby="more-detail-hint more-detail-error"
+                defaultValue={referralData['referral-reason']}
+                onChange={e => {
+                  setReferralData({ ...referralData, 'referral-reason': e.target.value });
+                }}
+                required
+                onInvalid={e => onInvalidField(e.target.id)}></textarea>
+            </div>
+            <div
+              className={`govuk-form-group govuk-!-padding-bottom-2 ${
+                validationError[`conversation-notes-${id}`] ? 'govuk-form-group--error' : ''
+              }`}>
+              <legend className="govuk-fieldset__legend govuk-fieldset__legend--s">
+                Notes on wider conversation(other needs, living situation, key information
+              </legend>
+              {validationError[`conversation-notes-${id}`] && (
+                <span
+                  id="more-detail-error"
+                  className={` ${
+                    validationError[`conversation-notes-${id}`] ? 'govuk-error-message' : ''
+                  }`}>
+                  <span className="govuk-visually-hidden">Error:</span> Enter more detail
+                </span>
+              )}
+              <textarea
+                className={`govuk-textarea ${
+                  validationError[`conversation-notes-${id}`] ? 'govuk-input--error' : ''
+                }`}
+                id={`conversation-notes-${id}`}
+                name="conversation-notes"
+                rows="5"
+                aria-describedby="more-detail-hint more-detail-error"
+                defaultValue={referralData['conversation-notes']}
+                onChange={e => {
+                  setReferralData({
+                    ...referralData,
+                    'conversation-notes': e.target.value
+                  });
+                }}
+                required
+                onInvalid={e => onInvalidField(e.target.id)}></textarea>
+            </div>
+            <div
+              className={`govuk-form-group govuk-!-padding-bottom-2 ${
+                validationError[`consent-${id}`] ? 'govuk-form-group--error' : ''
+              }`}>
+              <fieldset className="govuk-fieldset" aria-describedby="consent-hint">
+                <legend className="govuk-fieldset__legend govuk-fieldset__legend--s">
+                  The resident is happy for their information to be shared with third parties
+                </legend>
+                <div className="govuk-checkboxes govuk-checkboxes--inline">
+                  {validationError[`consent-${id}`] && (
+                    <span id={`consent-${id}-error`} className="govuk-error-message">
+                      <span className="govuk-visually-hidden">Error:</span>
+                      This is required in order to make the referral.
+                    </span>
+                  )}
+                  <div className="govuk-checkboxes__item">
+                    <input
+                      className="govuk-checkboxes__input"
+                      id={`consent-${id}`}
+                      name="consent"
+                      type="checkbox"
+                      value="true"
+                      onInvalid={e => onInvalidField(e.target.id)}
+                      required
+                    />
+                    <label
+                      className="govuk-label govuk-checkboxes__label"
+                      htmlFor={`consent-${id}`}>
+                      Yes
+                    </label>
+                  </div>
+                </div>
+              </fieldset>
+            </div>
+            <div className="govuk-form-group govuk-!-padding-bottom-2">
+              <fieldset className="govuk-fieldset">
+                <legend className="govuk-fieldset__legend govuk-fieldset__legend--s">
+                  How would the resident like to receive details about their referral to this
+                  service?
+                </legend>
+                <div id="resident-referral-hint" className="govuk-hint">
+                  Select all that apply.
+                </div>
+                <div className="govuk-checkboxes">
+                  <div className="govuk-checkboxes__item">
+                    <input
+                      className="govuk-checkboxes__input"
+                      id="resident-referral-sms"
+                      name="resident-referral-sms"
+                      type="checkbox"
+                      value={true}
+                    />
+                    <label
+                      className="govuk-label govuk-checkboxes__label"
+                      htmlFor="resident-referral-sms">
+                      By text
+                    </label>
+                  </div>
+                  <div className="govuk-checkboxes__item">
+                    <input
+                      className="govuk-checkboxes__input"
+                      id="resident-referral-email"
+                      name="resident-referral-email"
+                      type="checkbox"
+                      value={true}
+                    />
+                    <label
+                      className="govuk-label govuk-checkboxes__label"
+                      htmlFor="resident-referral-email">
+                      By email
+                    </label>
+                  </div>
+                </div>
+              </fieldset>
+            </div>
+            <div>
+              <div
+                className={`govuk-form-group govuk-!-padding-bottom-2 ${
+                  validationError[`referer-name-${id}`] ? 'govuk-form-group--error' : ''
+                }`}>
+                <label className="govuk-label inline" htmlFor={`referer-name-${id}`}>
+                  <legend className="govuk-fieldset__legend govuk-fieldset__legend--s">
+                    Your name
+                  </legend>
+                </label>
+                <span
+                  id="name-error"
+                  className="govuk-error-message"
+                  aria-describedby="input-name-error">
+                  <span hidden={!validationError[`referer-name-${id}`]} data-testid="name-error">
+                    Enter your name
+                  </span>
+                </span>
+                <input
+                  className={`govuk-input govuk-!-width-two-thirds ${
+                    validationError[`referer-name-${id}`] ? 'govuk-input--error' : ''
+                  }`}
+                  id={`referer-name-${id}`}
+                  name="referer-name"
+                  type="text"
+                  defaultValue={referrerData['referer-name']}
+                  onChange={e => {
+                    setReferrerData({ ...referrerData, 'referer-name': e.target.value });
+                  }}
+                  aria-describedby="refererName-hint"
+                  aria-describedby="refererName"
+                  onInvalid={e => onInvalidField(e.target.id)}
+                  required
+                />
+              </div>
+              <div
+                className={`govuk-form-group govuk-!-padding-bottom-2 ${
+                  validationError[`referer-organisation-${id}`] ? 'govuk-form-group--error' : ''
+                }`}>
+                <label className="govuk-label inline" htmlFor={`referer-organisation-${id}`}>
+                  <legend className="govuk-fieldset__legend govuk-fieldset__legend--s">
+                    Your organisation
+                  </legend>
+                </label>
+                <span
+                  id="name-error"
+                  className="govuk-error-message"
+                  aria-describedby="input-name-error">
+                  <span
+                    hidden={!validationError[`referer-organisation-${id}`]}
+                    data-testid="name-error">
+                    Enter your organisation
+                  </span>
+                </span>
+                <input
+                  className={`govuk-input govuk-!-width-two-thirds ${
+                    validationError[`referer-organisation-${id}`] ? 'govuk-input--error' : ''
+                  }`}
+                  id={`referer-organisation-${id}`}
+                  name="referer-organisation"
+                  type="text"
+                  defaultValue={referrerData['referer-organisation']}
+                  onChange={e => {
+                    setReferrerData({
+                      ...referrerData,
+                      'referer-organisation': e.target.value
+                    });
+                  }}
+                  aria-describedby="refererorganisation-hint"
+                  aria-describedby="refererorganisation"
+                  onInvalid={e => onInvalidField(e.target.id)}
+                  required
+                />
+              </div>
+              <div
+                className={`govuk-form-group govuk-!-padding-bottom-2 ${
+                  validationError[`referer-email-${id}`] ? 'govuk-form-group--error' : ''
+                }`}>
+                <label className="govuk-label inline" htmlFor={`referer-email-${id}`}>
+                  <legend className="govuk-fieldset__legend govuk-fieldset__legend--s">
+                    Your email
+                  </legend>
+                </label>
+                <span
+                  id="name-error"
+                  className="govuk-error-message"
+                  aria-describedby="input-name-error">
+                  <span hidden={!validationError[`referer-email-${id}`]} data-testid="name-error">
+                    Enter your email
+                  </span>
+                </span>
+                <input
+                  className={`govuk-input govuk-!-width-two-thirds ${
+                    validationError[`referer-email-${id}`] ? 'govuk-input--error' : ''
+                  }`}
+                  id={`referer-email-${id}`}
+                  name="referer-email"
+                  type="email"
+                  defaultValue={referrerData['referer-email']}
+                  onChange={e => {
+                    setReferrerData({ ...referrerData, 'referer-email': e.target.value });
+                  }}
+                  aria-describedby="refererorganisation-hint"
+                  aria-describedby="refererorganisation"
+                  onInvalid={e => onInvalidField(e.target.id)}
+                  required
+                />
+                <input value={id} name="service-id" hidden />
+              </div>
+            </div>
+            <div className="govuk-form-group govuk-!-padding-bottom-2">
+              <div className="govuk-!-margin-top-4">
+                <input type="submit" id={`submit-${id}`} className="govuk-button" name="Submit" />
+              </div>
+            </div>
+          </div>
+        </form>
+      }
+    </div>
+  );
+};
+
+export default ReferralForm;
