+++ components/Feature/VulnerabilitiesGrid/ResourceCard/index.js
@@ -8,7 +8,7 @@ import {
   SERVICE_CLICK_WEBSITE,
   VIEW_SUMMARY_EMAIL_CLICKED
 } from 'lib/utils/analyticsConstants';
-import ResidentDetailsForm from 'components/Feature/ResidentDetailsForm';
+import ReferralForm from 'components/Feature/ReferralForm';
 
 const ResourceCard = ({
   id,
@@ -82,14 +82,14 @@ const ResourceCard = ({
     });
   };
 
-  const onInvalidField = value => {
-    Object.values(validationError).every(k => {
-      if (value != k) {
-        validationError[k] = false;
-      }
-    });
-    setValidationError({ [value]: true, ...validationError });
-  };
+  // const onInvalidField = value => {
+  //   Object.values(validationError).every(k => {
+  //     if (value != k) {
+  //       validationError[k] = false;
+  //     }
+  //   });
+  //   setValidationError({ [value]: true, ...validationError });
+  // };
 
   const fullDescription = serviceDescription + ' ' + description;
   const first = fullDescription?.substring(0, 250);
@@ -220,7 +220,7 @@ const ResourceCard = ({
                   openReferralForm.categoryName == categoryName &&
                   !referralCompletion[id] && (
                     <div id={`referral-${id}-${categoryId}-form`} className={css['referral-form']}>
-                      <ResidentDetailsForm
+                      <ReferralForm
                         residentInfoCallback={residentInfoCallback}
                         showResidentForm={showResidentForm}
                         setShowResidentForm={setShowResidentForm}
@@ -232,347 +232,18 @@ const ResourceCard = ({
                         updateEmailBody={updateEmailBody}
                         setEmailBody={setEmailBody}
                         referrerData={referrerData}
+                        id={id}
+                        email={email}
+                        name={name}
+                        description={description}
+                        websites={websites}
+                        address={address}
+                        postcode={postcode}
+                        tags={tags}
+                        telephone={telephone}
+                        referralContact={referralContact}
+                        referralData={referralData}
                       />
-                      <input
-                        form="resident-details"
-                        id={`service-name-${id}`}
-                        name="service-name"
-                        value={name}
-                        type="text"
-                        hidden
-                      />
-                      <input
-                        form="resident-details"
-                        id={`service-contact-email-${id}`}
-                        name="service-contact-email"
-                        value={email}
-                        type="text"
-                        hidden
-                      />
-                      <input
-                        form="resident-details"
-                        id={`service-contact-phone-${id}`}
-                        name="service-contact-phone"
-                        value={telephone}
-                        type="text"
-                        hidden
-                      />
-                      <input
-                        form="resident-details"
-                        id={`service-referral-email-${id}`}
-                        name="service-referral-email"
-                        value={referralContact}
-                        type="text"
-                        hidden
-                      />
-                      <input
-                        form="resident-details"
-                        id={`service-address-${id}`}
-                        name="service-address"
-                        value={address}
-                        type="text"
-                        hidden
-                      />
-                      <input
-                        form="resident-details"
-                        id={`service-websites-${id}`}
-                        name="service-websites"
-                        value={websites.join(' ')}
-                        type="text"
-                        hidden
-                      />
-                      <input
-                        form="resident-details"
-                        id={`service-description-${id}`}
-                        name="service-description"
-                        value={description}
-                        type="text"
-                        hidden
-                      />
-                      <div
-                        className={`govuk-form-group govuk-!-padding-bottom-2 govuk-!-padding-top-4 ${
-                          validationError[`referral-reason-${id}`] ? 'govuk-form-group--error' : ''
-                        }`}>
-                        <legend className="govuk-fieldset__legend govuk-fieldset__legend--s">
-                          Reason for referral, please give as much detail as possible
-                        </legend>
-                        {validationError[`referral-reason-${id}`] && (
-                          <span
-                            id="more-detail-error"
-                            className={` ${
-                              validationError[`referral-reason-${id}`] ? 'govuk-error-message' : ''
-                            }`}>
-                            <span className="govuk-visually-hidden">Error:</span> Enter more detail
-                          </span>
-                        )}
-                        <textarea
-                          form="resident-details"
-                          className={`govuk-textarea ${
-                            validationError[`referral-reason-${id}`] ? 'govuk-input--error' : ''
-                          }`}
-                          id={`referral-reason-${id}`}
-                          name="referral-reason"
-                          rows="5"
-                          aria-describedby="more-detail-hint more-detail-error"
-                          defaultValue={referralData['referral-reason']}
-                          onChange={e => {
-                            setReferralData({ ...referralData, 'referral-reason': e.target.value });
-                          }}
-                          required
-                          onInvalid={e => onInvalidField(e.target.id)}></textarea>
-                      </div>
-                      <div
-                        className={`govuk-form-group govuk-!-padding-bottom-2 ${
-                          validationError[`conversation-notes-${id}`]
-                            ? 'govuk-form-group--error'
-                            : ''
-                        }`}>
-                        <legend className="govuk-fieldset__legend govuk-fieldset__legend--s">
-                          Notes on wider conversation(other needs, living situation, key information
-                        </legend>
-                        {validationError[`conversation-notes-${id}`] && (
-                          <span
-                            id="more-detail-error"
-                            className={` ${
-                              validationError[`conversation-notes-${id}`]
-                                ? 'govuk-error-message'
-                                : ''
-                            }`}>
-                            <span className="govuk-visually-hidden">Error:</span> Enter more detail
-                          </span>
-                        )}
-                        <textarea
-                          form="resident-details"
-                          className={`govuk-textarea ${
-                            validationError[`conversation-notes-${id}`] ? 'govuk-input--error' : ''
-                          }`}
-                          id={`conversation-notes-${id}`}
-                          name="conversation-notes"
-                          rows="5"
-                          aria-describedby="more-detail-hint more-detail-error"
-                          defaultValue={referralData['conversation-notes']}
-                          onChange={e => {
-                            setReferralData({
-                              ...referralData,
-                              'conversation-notes': e.target.value
-                            });
-                          }}
-                          required
-                          onInvalid={e => onInvalidField(e.target.id)}></textarea>
-                      </div>
-                      <div
-                        className={`govuk-form-group govuk-!-padding-bottom-2 ${
-                          validationError[`consent-${id}`] ? 'govuk-form-group--error' : ''
-                        }`}>
-                        <fieldset className="govuk-fieldset" aria-describedby="consent-hint">
-                          <legend className="govuk-fieldset__legend govuk-fieldset__legend--s">
-                            The resident is happy for their information to be shared with third
-                            parties
-                          </legend>
-                          <div className="govuk-checkboxes govuk-checkboxes--inline">
-                            {validationError[`consent-${id}`] && (
-                              <span id={`consent-${id}-error`} className="govuk-error-message">
-                                <span className="govuk-visually-hidden">Error:</span>
-                                This is required in order to make the referral.
-                              </span>
-                            )}
-                            <div className="govuk-checkboxes__item">
-                              <input
-                                form="resident-details"
-                                className="govuk-checkboxes__input"
-                                id={`consent-${id}`}
-                                name="consent"
-                                type="checkbox"
-                                value="true"
-                                onInvalid={e => onInvalidField(e.target.id)}
-                                required
-                              />
-                              <label
-                                className="govuk-label govuk-checkboxes__label"
-                                htmlFor={`consent-${id}`}>
-                                Yes
-                              </label>
-                            </div>
-                          </div>
-                        </fieldset>
-                      </div>
-                      <div className="govuk-form-group govuk-!-padding-bottom-2">
-                        <fieldset className="govuk-fieldset">
-                          <legend className="govuk-fieldset__legend govuk-fieldset__legend--s">
-                            How would the resident like to receive details about their referral to
-                            this service?
-                          </legend>
-                          <div id="resident-referral-hint" className="govuk-hint">
-                            Select all that apply.
-                          </div>
-                          <div className="govuk-checkboxes">
-                            <div className="govuk-checkboxes__item">
-                              <input
-                                className="govuk-checkboxes__input"
-                                id="resident-referral-sms"
-                                name="resident-referral-sms"
-                                type="checkbox"
-                                form="resident-details"
-                                value={true}
-                              />
-                              <label
-                                className="govuk-label govuk-checkboxes__label"
-                                htmlFor="resident-referral-sms">
-                                By text
-                              </label>
-                            </div>
-                            <div className="govuk-checkboxes__item">
-                              <input
-                                className="govuk-checkboxes__input"
-                                id="resident-referral-email"
-                                name="resident-referral-email"
-                                type="checkbox"
-                                form="resident-details"
-                                value={true}
-                              />
-                              <label
-                                className="govuk-label govuk-checkboxes__label"
-                                htmlFor="resident-referral-email">
-                                By email
-                              </label>
-                            </div>
-                          </div>
-                        </fieldset>
-                      </div>
-                      <div>
-                        <div
-                          className={`govuk-form-group govuk-!-padding-bottom-2 ${
-                            validationError[`referer-name-${id}`] ? 'govuk-form-group--error' : ''
-                          }`}>
-                          <label className="govuk-label inline" htmlFor={`referer-name-${id}`}>
-                            <legend className="govuk-fieldset__legend govuk-fieldset__legend--s">
-                              Your name
-                            </legend>
-                          </label>
-                          <span
-                            id="name-error"
-                            className="govuk-error-message"
-                            aria-describedby="input-name-error">
-                            <span
-                              hidden={!validationError[`referer-name-${id}`]}
-                              data-testid="name-error">
-                              Enter your name
-                            </span>
-                          </span>
-                          <input
-                            form="resident-details"
-                            className={`govuk-input govuk-!-width-two-thirds ${
-                              validationError[`referer-name-${id}`] ? 'govuk-input--error' : ''
-                            }`}
-                            id={`referer-name-${id}`}
-                            name="referer-name"
-                            type="text"
-                            defaultValue={referrerData['referer-name']}
-                            onChange={e => {
-                              setReferrerData({ ...referrerData, 'referer-name': e.target.value });
-                            }}
-                            aria-describedby="refererName-hint"
-                            aria-describedby="refererName"
-                            onInvalid={e => onInvalidField(e.target.id)}
-                            required
-                          />
-                        </div>
-                        <div
-                          className={`govuk-form-group govuk-!-padding-bottom-2 ${
-                            validationError[`referer-organisation-${id}`]
-                              ? 'govuk-form-group--error'
-                              : ''
-                          }`}>
-                          <label
-                            className="govuk-label inline"
-                            htmlFor={`referer-organisation-${id}`}>
-                            <legend className="govuk-fieldset__legend govuk-fieldset__legend--s">
-                              Your organisation
-                            </legend>
-                          </label>
-                          <span
-                            id="name-error"
-                            className="govuk-error-message"
-                            aria-describedby="input-name-error">
-                            <span
-                              hidden={!validationError[`referer-organisation-${id}`]}
-                              data-testid="name-error">
-                              Enter your organisation
-                            </span>
-                          </span>
-                          <input
-                            form="resident-details"
-                            className={`govuk-input govuk-!-width-two-thirds ${
-                              validationError[`referer-organisation-${id}`]
-                                ? 'govuk-input--error'
-                                : ''
-                            }`}
-                            id={`referer-organisation-${id}`}
-                            name="referer-organisation"
-                            type="text"
-                            defaultValue={referrerData['referer-organisation']}
-                            onChange={e => {
-                              setReferrerData({
-                                ...referrerData,
-                                'referer-organisation': e.target.value
-                              });
-                            }}
-                            aria-describedby="refererorganisation-hint"
-                            aria-describedby="refererorganisation"
-                            onInvalid={e => onInvalidField(e.target.id)}
-                            required
-                          />
-                        </div>
-                        <div
-                          className={`govuk-form-group govuk-!-padding-bottom-2 ${
-                            validationError[`referer-email-${id}`] ? 'govuk-form-group--error' : ''
-                          }`}>
-                          <label className="govuk-label inline" htmlFor={`referer-email-${id}`}>
-                            <legend className="govuk-fieldset__legend govuk-fieldset__legend--s">
-                              Your email
-                            </legend>
-                          </label>
-                          <span
-                            id="name-error"
-                            className="govuk-error-message"
-                            aria-describedby="input-name-error">
-                            <span
-                              hidden={!validationError[`referer-email-${id}`]}
-                              data-testid="name-error">
-                              Enter your email
-                            </span>
-                          </span>
-                          <input
-                            form="resident-details"
-                            className={`govuk-input govuk-!-width-two-thirds ${
-                              validationError[`referer-email-${id}`] ? 'govuk-input--error' : ''
-                            }`}
-                            id={`referer-email-${id}`}
-                            name="referer-email"
-                            type="email"
-                            defaultValue={referrerData['referer-email']}
-                            onChange={e => {
-                              setReferrerData({ ...referrerData, 'referer-email': e.target.value });
-                            }}
-                            aria-describedby="refererorganisation-hint"
-                            aria-describedby="refererorganisation"
-                            onInvalid={e => onInvalidField(e.target.id)}
-                            required
-                          />
-                          <input form="resident-details" value={id} name="service-id" hidden />
-                        </div>
-                      </div>
-                      <div className="govuk-form-group govuk-!-padding-bottom-2">
-                        <div className="govuk-!-margin-top-4">
-                          <input
-                            type="submit"
-                            id={`submit-${id}`}
-                            className="govuk-button"
-                            name="Submit"
-                            form="resident-details"
-                          />
-                        </div>
-                      </div>
                     </div>
                   )}
               </span>
